Revert "Flakyness: backup/restore - The 'internal' tenant contains error(s) or warning(s)"
Because the restore command of the CLI stop ZK, then this process in the
CI is flaky and usually succeed by ZK is still up but in the process of
being stopped. So sometime it fails (seen in the CI).

So let's revert and think to add this to the restore command instead.

This reverts commit a5715d12062de656741d47768f66c3583c1ca058.

Change-Id: I93cb1836f9ed6f60546a88139ea6bf2ea895484a