CI: Test executor graceful stop
This change adds a scenario in CI to validate that shutting down
executors via kubernetes orchestration is done gracefully, ie
the executor(s) will finish any scheduled job before stopping.
This is useful to ensure minimal interruption of service during
maintenance (resources scaling or upgrades).

The scenario's steps are:

- Create a long lived job and run it in the check pipeline
- Scale down executor while the job is running
- Add a shorter job, ensure it is queued but not running
- Trigger first job completion
- Wait until executor pod is terminated
- Ensure long lived job completed
- Scale up executor
- Ensure second job completes

Add a helper role to create a file on the log server.
Add a demo job (unused by default) that waits for a
file to be present on the logserver to complete.

Also ensure logserver contents, repos are fetched in post
during the minikube job.

Change-Id: Ibbe16270293b98a0f1608c42114abec884ec8cfd