Fix nodeAffinity missing weight (defaulting to 0)
The 0 value is forbidden (seen on OpenShift and Minikube(k8s)). This
patch set a weight to 100 which is the stronger value to expected the
re-scheduling on the expected node.

```
Warning  FailedCreate         57s (x15 over 3m1s)  statefulset-controller  create Pod logserver-0 in StatefulSet logserver failed error: Pod "logserver-0" is invalid: spec.affinity.nodeAffinity.preferredDuringSchedulingIgnoredDuringExecution[0].weight: Invalid value: 0: must be in the range 1-100
```

Change-Id: I63fe48be1ead00abae8bb76acac8ded2334d1c5b