Enable fail-fast for Zuul gate pipeline
- Enable the fail-fast setting [1] at the gate
pipeline.

With this enabled, Zuul will immediately cancel all
remaining jobs in a pipeline when a single job fails. This helps save
resources and provides quicker feedback on failures, particularly in
resource-constrained environments.

We don't enable it at check pipeline, in order to provide a full
comprehensive report of what should be addressed.

[1] https://zuul-ci.org/docs/zuul/latest/config/project.html#attr-project.%3Cpipeline%3E.fail-fast

Change-Id: If7976847de39705211661bbc5b875b73595f497d