Enhance --dry-run to show specific reasons for updates
- Improves the output of the `deploy --dry-run` command by providing
detailed reasons for why a resource would be updated.

Instead of a generic "Spec template has changed" message, the operator
now compares the annotations of the current and desired resources and
lists the specific annotations that were added, removed, or changed.
This is achieved via a new `compareAnnotations` helper function.

- In addition, the CreateR, UpdateR, and DeleteR methods now pass
the DryRunAll option to the Kubernetes API when the operator is in
dry-run mode. This ensures that all changes are validated by the
server without being persisted.

- Move the Changelog entry to the right place.

This completes the implementation of the functionality requested in the
Jira ticket [1].

[1] RHOSZUUL-2753

Change-Id: Idd4982b59a75ccdd578eaa916c346a8d6723a417