nodeAffinity: handle case when nodes cannot be figured out
When a deployment or stateful set has 0 replica, it is still possible
to list its pods (returns an empty list), but we obviously can't figure
out the nodes used by ... nothing.

Handle this case in the nodeAffinity support logic. Log the issue
and suggest to re-run a reconciliation to effectively set the
nodeAffinity from existing pods.

Change-Id: Ibfeaeeb8d8883bad9a0056c17c4112aaf0d0f229