zookeeper: prepare statefulset, other components to support multiple replicas
Establish the ground work for zookeeper to be deployed as an ensemble
of 3 replicas:

- fix init script to configure the hosts in the quorum properly
- Generate and update proper client certs, server certs
- introduce new CRD parameter to define external ZK host(s) as
  a list
- Add client port to headless service, so that clients can connect
  to replicas directly as intended
- Configure zuul, nodepool components to access directly the replicas

Since we want to ensure a proper upgrade path to the 3 replicas setup
in operator mode, the next release will keep zookeeper deployed as a
single replica statefulset. The bump to 3 replicas will be enforced in a later
release.

Change-Id: I669bd5812b8d5bfc8928307350118dcb8695434c